Now support for Custom Webinterface(Code not public YET)
Updated function to disable certain messages(Like they where TO MUTCH spam)

Fixed a lot of buggs with pausing while playing when someone joines.

Changed everything to DUser instead of using the building ctx(This was because of the webinterface(And this is way nicer) )

Added Webinterface & Remote control

And ofcourse A LOT of other stuff.